# docs/swagger.yaml
basePath: /v1
definitions:
  models.CaptchaResponse:
    properties:
      captchaId:
        description: 验证码Id
        type: string
      imageUrl:
        description: 验证码图片url
        type: string
    type: object
...
  models.Tenant:
    properties:
      cctasp:
        description: 云计算技术和服务提供商
        type: string
      created_at:
        description: 创建时间戳
        type: integer
      disabled_is:
        description: 1:禁用当前记录
        type: integer
      due_at:
        description: 承租到期日期
        type: string
      id:
        description: 承租户id
        type: integer
      start_at:
        description: 承租起始日期
        type: string
      status:
        description: 状态
        type: string
      tenant_des:
        description: 承租户描述
        type: string
      tenant_id:
        description: 承租户uuid
        type: string
      tenant_name:
        description: 承租户名
        type: string
      updated_at:
        description: 创建时间戳
        type: integer
    required:
    - tenant_name
    type: object
  models.Token:
    properties:
      access_token:
        description: token uuid
        type: string
      expires_in:
        description: 存活时间3600秒 1时间
        type: integer
      id:
        description: token id
        type: integer
      refresh_token:
        description: 刷新token uuid
        type: string
      scope:
        description: 范围
        type: string
      token_type:
        description: token类型
        type: string
    type: object
  models.User:
    properties:
      created_at:
        description: 创建时间戳
        type: integer
      email:
        description: 用户邮箱
        type: string
      first_name:
        description: 用户名
        type: string
      gender:
        description: 性别
        type: string
      id:
        description: 用户id
        type: integer
      last_name:
        description: 用户名
        type: string
      mobile_phone:
        description: 移动电话
        type: string
      password:
        description: 用户密码
        type: string
      tenant_id:
        description: tenant_id
        type: string
      updated_at:
        description: 创建时间戳
        type: integer
      user_name:
        description: 用户名
        type: string
      uuid:
        description: 用户uuid
        type: string
    required:
    - email
    - mobile_phone
    - password
    - tenant_id
    - user_name
    type: object
...
  swag.ResultContChangePass:
    properties:
      data:
        $ref: '#/definitions/models.User'
        description: 返回的数据
      msg:
        description: 报错信息
        example: 错误
        type: string
      statusCode:
        description: 代码,0:成功，非0:失败
        example: 0
        type: integer
    type: object
  swag.ResultContCreatedTenant:
    properties:
      data:
        $ref: '#/definitions/models.Tenant'
        description: 返回的数据
      msg:
        description: 报错信息
        example: 错误
        type: string
      statusCode:
        description: 代码,0:成功，非0:失败
        example: 0
        type: integer
    type: object
  swag.ResultContDeleteTenantById:
    properties:
      data:
        $ref: '#/definitions/models.Tenant'
        description: 返回的数据
      msg:
        description: 报错信息
        example: 错误
        type: string
      statusCode:
        description: 代码,0:成功，非0:失败
        example: 0
        type: integer
    type: object
  swag.ResultContGetCaptchaImage:
    properties:
      data:
        $ref: '#/definitions/models.CaptchaResponse'
        description: 返回的数据
      msg:
        description: 报错信息
        example: 错误
        type: string
      statusCode:
        description: 代码,0:成功，非0:失败
        example: 0
        type: integer
    type: object
  swag.ResultContGetTenantById:
    properties:
      data:
        $ref: '#/definitions/models.Tenant'
        description: 返回的数据
      msg:
        description: 报错信息
        example: 错误
        type: string
      statusCode:
        description: 代码,0:成功，非0:失败
        example: 0
        type: integer
    type: object
  swag.ResultContGetTenantByName:
    properties:
      data:
        $ref: '#/definitions/models.Tenant'
        description: 返回的数据
      msg:
        description: 报错信息
        example: 错误
        type: string
      statusCode:
        description: 代码,0:成功，非0:失败
        example: 0
        type: integer
    type: object
  swag.ResultContGetTenantList:
    properties:
      data:
        $ref: '#/definitions/models.Tenant'
        description: 返回的数据
      msg:
        description: 报错信息
        example: 错误
        type: string
      statusCode:
        description: 代码,0:成功，非0:失败
        example: 0
        type: integer
    type: object
...
  swag.ResultContSingIn:
    properties:
      data:
        $ref: '#/definitions/models.User'
        description: 返回的数据
      msg:
        description: 报错信息
        example: 错误
        type: string
      statusCode:
        description: 代码,0:成功，非0:失败
        example: 0
        type: integer
    type: object
  swag.ResultContSingUp:
    properties:
      data:
        $ref: '#/definitions/models.User'
        description: 返回的数据
      msg:
        description: 报错信息
        example: 错误
        type: string
      statusCode:
        description: 代码,0:成功，非0:失败
        example: 0
        type: integer
    type: object
  swag.ResultContTest:
    properties:
      data:
        $ref: '#/definitions/models.User'
        description: 返回的数据
      msg:
        description: 报错信息
        example: 错误
        type: string
      statusCode:
        description: 代码,0:成功，非0:失败
        example: 0
        type: integer
    type: object
  swag.ResultContToken:
    properties:
      data:
        $ref: '#/definitions/models.Token'
        description: 返回的数据
      msg:
        description: 报错信息
        example: 错误
        type: string
      statusCode:
        description: 代码,0:成功，非0:失败
        example: 0
        type: integer
    type: object
  swag.ResultContUpdateTenant:
    properties:
      data:
        $ref: '#/definitions/models.Tenant'
        description: 返回的数据
      msg:
        description: 报错信息
        example: 错误
        type: string
      statusCode:
        description: 代码,0:成功，非0:失败
        example: 0
        type: integer
    type: object
  swag.ResultContUserByEmail:
    properties:
      data:
        $ref: '#/definitions/models.User'
        description: 返回的数据
      msg:
        description: 报错信息
        example: 错误
        type: string
      statusCode:
        description: 代码,0:成功，非0:失败
        example: 0
        type: integer
    type: object
  swag.ResultContUserById:
    properties:
      data:
        $ref: '#/definitions/models.User'
        description: 返回的数据
      msg:
        description: 报错信息
        example: 错误
        type: string
      statusCode:
        description: 代码,0:成功，非0:失败
        example: 0
        type: integer
    type: object
  swag.ResultContUserByName:
    properties:
      data:
        $ref: '#/definitions/models.User'
        description: 返回的数据
      msg:
        description: 报错信息
        example: 错误
        type: string
      statusCode:
        description: 代码,0:成功，非0:失败
        example: 0
        type: integer
    type: object
  swag.ResultContUserByTenantId:
    properties:
      data:
        $ref: '#/definitions/models.User'
        description: 返回的数据
      msg:
        description: 报错信息
        example: 错误
        type: string
      statusCode:
        description: 代码,0:成功，非0:失败
        example: 0
        type: integer
    type: object
  swag.ResultContVerifyCaptcha:
    properties:
      data:
        $ref: '#/definitions/models.CaptchaResponse'
        description: 返回的数据
      msg:
        description: 报错信息
        example: 错误
        type: string
      statusCode:
        description: 代码,0:成功，非0:失败
        example: 0
        type: integer
    type: object
host: localhost:9090
info:
  contact: {}
  description: XXXX项目描述
  title: XXXX项目名称-Swagger Example API
  version: "1.0"
paths:
...
  /captcha/{captchaId}:
    get:
      description: 获取验证码 GetCaptcha
      parameters:
      - description: 验证码Id
        in: path
        name: captchaId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContGetCaptchaImage'
      summary: 获取验证码图片
  /captcha/verify/{captchaId}/{value}:
    get:
      description: 获取验证码 GetCaptcha
      parameters:
      - description: 验证码Id
        in: path
        name: captchaId
        required: true
        type: integer
      - description: 验证码value
        in: path
        name: value
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContVerifyCaptcha'
      summary: 验证 验证码码图
  /changepass:
    put:
      description: 输入用户名及密登录
      parameters:
      - description: 用户邮箱
        in: path
        name: email
        required: true
        type: string
      - description: 旧密码
        in: path
        name: oldpassword
        required: true
        type: string
      - description: 新密码
        in: path
        name: newpassword
        required: true
        type: string
      - description: 确认新密码
        in: path
        name: confirmpassword
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContChangePass'
      summary: 修改密码
...
  /login:
    post:
      description: 输入用户名及密登录
      parameters:
      - description: 登录名
        in: path
        name: username
        required: true
        type: string
      - description: 登录密码
        in: path
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContSingIn'
      summary: 系统超级管理登录
  /singin:
    post:
      description: 输入用户名及密登录
      parameters:
      - description: 承租
        in: path
        name: tenant
        required: true
        type: string
      - description: 登录名
        in: path
        name: username
        required: true
        type: string
      - description: 登录密码
        in: path
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContSingIn'
      summary: 系统登录
  /singup:
    post:
      description: 输入用户名及密登录
      parameters:
      - description: 承租户tenant_id
        in: path
        name: tenant_id
        required: true
        type: integer
      - description: 用户邮箱
        in: path
        name: email
        required: true
        type: string
      - description: 登录名
        in: path
        name: username
        required: true
        type: string
      - description: 登录密码
        in: path
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContSingUp'
      summary: 注册用户
  /tenant:
    get:
      description: 显示承租户详情
      parameters:
      - description: 承租pageNo
        in: path
        name: pageNo
        required: true
        type: string
      - description: 承租pageSize
        in: path
        name: pageSize
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContGetTenantList'
      security:
      - ApiKeyAuth: []
      summary: 获取所有承租户详情
    post:
      description: 输入承租户名及描述
      parameters:
      - description: 承租户名
        in: path
        name: tenant_name
        required: true
        type: string
      - description: 承租户描述
        in: path
        name: tenant_des
        type: string
      - description: 云计算技术和服务提供商
        in: path
        name: cctasp
        type: string
      - description: 承租起始日期
        in: path
        name: start_at
        type: string
      - description: 承租到期日期
        in: path
        name: due_at
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContCreatedTenant'
      security:
      - ApiKeyAuth: []
      summary: 创建承租户
    put:
      description: 输入承租户名及描述
      parameters:
      - description: 承租户tenant_id
        in: path
        name: tenant_id
        required: true
        type: integer
      - description: 承租户名
        in: path
        name: tenant_name
        required: true
        type: string
      - description: 承租户描述
        in: path
        name: tenant_des
        type: string
      - description: 云计算技术和服务提供商
        in: path
        name: cctasp
        type: string
      - description: 承租起始日期
        in: path
        name: start_at
        type: string
      - description: 承租到期日期
        in: path
        name: due_at
        type: string
      - description: 1/true禁用当前记录
        in: path
        name: disabled_is
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContUpdateTenant'
      security:
      - ApiKeyAuth: []
      summary: 修改承租户
  /tenant/{tenant_id}:
    delete:
      description: 删除承租户
      parameters:
      - description: 承租id
        in: path
        name: tenant_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContDeleteTenantById'
      security:
      - ApiKeyAuth: []
      summary: 删除承租户
    get:
      description: 显示承租户详情
      parameters:
      - description: 承租id
        in: path
        name: tenant_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContGetTenantById'
      security:
      - ApiKeyAuth: []
      summary: 用tenant_id获取承租户详情
  /tenant/name/{tenant_name}:
    get:
      description: 显示承租户详情
      parameters:
      - description: 承租tenant_name
        in: path
        name: tenant_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContGetTenantByName'
      security:
      - ApiKeyAuth: []
      summary: 用tenant_name获取承租户详情
  /test:
    get:
      description: Test===
      parameters:
      - description: check_in
        in: path
        name: check_in
        required: true
        type: string
      - description: check_out
        in: path
        name: check_out
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContTest'
      summary: Test
  /token:
    get:
      description: 获取Token
      parameters:
      - description: 承租
        in: path
        name: tenant
        required: true
        type: string
      - description: 登录名
        in: path
        name: user
        required: true
        type: string
      - description: 登录密码
        in: path
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContToken'
      summary: 取Token
    post:
      description: 获取Token
      parameters:
      - description: 承租
        in: path
        name: tenant
        required: true
        type: string
      - description: 登录名
        in: path
        name: name
        required: true
        type: string
      - description: 登录密码
        in: path
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContToken'
      summary: 生成Token
  /user/{id}:
    get:
      description: 显示用户的详情
      parameters:
      - description: 用户id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContUserById'
      summary: 用户详情
  /user/email/{email}:
    get:
      description: 显示用户的详情
      parameters:
      - description: 用户email
        in: path
        name: email
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContUserByEmail'
      summary: 用户详情
  /user/name/{name}:
    get:
      description: 显示用户的详情
      parameters:
      - description: 用户name
        in: path
        name: name
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContUserByName'
      summary: 用户详情
  /user/tenant/{tenant_id}:
    get:
      description: 显示用户列表
      parameters:
      - description: 分类id
        in: path
        name: tenant_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swag.ResultContUserByTenantId'
      summary: 用户列表
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: token
    type: apiKey
swagger: "2.0"